+++ application/common/mail/method-verify.html.php
@@ -22,7 +22,7 @@ $verifyLink = $passwordProfileUrl . '/password/recovery/' . $uid . '/verify/' . 
 <p>
     Someone recently requested to add this email address, <?= yHtml::encode($toAddress); ?>,
     as a method for verifying themselves should they need to reset their
-    <?= yHtml::encode($idpDisplayName); ?> account password. If this was you, you may click
+    <?= yHtml::encode($idpDisplayName); ?> Identity account password. If this was you, you may click
     <?= yHtml::a(yHtml::encode($verifyLink), $verifyLink) ?>
     to add it to your account.
 </p>