+++ application/common/mail/mfa-manager.html.php
@@ -16,7 +16,7 @@ use yii\helpers\Html as yHtml;
 
 <p>
     <?= yHtml::encode($displayName) ?> has requested your assistance in accessing their
-    <?= yHtml::encode($idpDisplayName) ?> account. This email contains a backup code they need
+    <?= yHtml::encode($idpDisplayName) ?> Identity account. This email contains a backup code they need
     to proceed with login. Please contact them directly to ensure that you are only providing
     the following backup code to them and not to someone else.
 </p>