+++ application/common/models/EmailLog.php
@@ -24,8 +24,12 @@ class EmailLog extends EmailLogBase
     const MESSAGE_TYPE_MFA_ENABLED = 'mfa-enabled';
     const MESSAGE_TYPE_MFA_DISABLED = 'mfa-disabled';
     const MESSAGE_TYPE_METHOD_VERIFY = 'method-verify';
+    const MESSAGE_TYPE_METHOD_REMINDER = 'method-reminder';
+    const MESSAGE_TYPE_METHOD_PURGED = 'method-purged';
     const MESSAGE_TYPE_MFA_MANAGER = 'mfa-manager';
     const MESSAGE_TYPE_MFA_MANAGER_HELP = 'mfa-manager-help';
+    const MESSAGE_TYPE_PASSWORD_EXPIRING = 'password-expiring';
+    const MESSAGE_TYPE_PASSWORD_EXPIRED = 'password-expired';
 
     /**
      * @inheritdoc
@@ -52,8 +56,12 @@ class EmailLog extends EmailLogBase
             self::MESSAGE_TYPE_MFA_ENABLED,
             self::MESSAGE_TYPE_MFA_DISABLED,
             self::MESSAGE_TYPE_METHOD_VERIFY,
+            self::MESSAGE_TYPE_METHOD_REMINDER,
+            self::MESSAGE_TYPE_METHOD_PURGED,
             self::MESSAGE_TYPE_MFA_MANAGER,
             self::MESSAGE_TYPE_MFA_MANAGER_HELP,
+            self::MESSAGE_TYPE_PASSWORD_EXPIRING,
+            self::MESSAGE_TYPE_PASSWORD_EXPIRED,
         ];
     }